# Conflicts:
#	Source/ACE.Server/Command/Handlers/DeveloperCommands.cs
#	Source/ACE.Server/Command/Handlers/PlayerCommands.cs
#	Source/ACE.Server/Factories/LootGenerationFactory.cs
#	Source/ACE.Server/Factories/LootGenerationFactory_Clothing.cs
#	Source/ACE.Server/Factories/LootGenerationFactory_Jewelry.cs
#	Source/ACE.Server/Factories/LootGenerationFactory_Magic.cs
#	Source/ACE.Server/Managers/PropertyManager.cs
#	Source/ACE.Server/Network/NetworkSession.cs
#	Source/ACE.Server/WorldObjects/Player_Magic.cs
#	Source/ACE.Server/WorldObjects/Player_Xp.cs

+++ Source/ACE.Server/Managers/RecipeManager.cs
@@ -577,7 +577,7 @@ namespace ACE.Server.Managers
 
                     // change wield requirement: missile defense -> melee defense
                     target.WieldSkillType = (int)Skill.MeleeDefense;
-                    target.ItemSkillLimit = (int)Skill.MeleeDefense;      // recipe requirements check for this field
+                    target.ItemSkillLimit = Skill.MeleeDefense;      // recipe requirements check for this field
 
                     // increase the wield difficulty
                     if (target.WieldDifficulty != null)
@@ -605,7 +605,7 @@ namespace ACE.Server.Managers
 
                     // change wield requirement: melee defense -> missile defense
                     target.WieldSkillType = (int)Skill.MissileDefense;
-                    target.ItemSkillLimit = (int)Skill.MissileDefense;      // recipe requirements check for this field
+                    target.ItemSkillLimit = Skill.MissileDefense;      // recipe requirements check for this field
 
                     // decrease the wield difficulty
                     if (target.WieldDifficulty != null)
